# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OMPay",
        "description": "API pour la gestion des comptes bancaires OMPay",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/marchands": {
            "get": {
                "tags": [
                    "Marchands"
                ],
                "summary": "Liste des marchands",
                "operationId": "67a7ea130af1d8169d488061982b0827",
                "responses": {
                    "200": {
                        "description": "Liste des marchands récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Marchand"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Marchands"
                ],
                "summary": "Créer un nouveau marchand",
                "operationId": "99b56cf99f34bdbf3d89ce96ecd8b7e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Marchand"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Marchand créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marchand"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/marchands/{id}": {
            "get": {
                "tags": [
                    "Marchands"
                ],
                "summary": "Afficher un marchand spécifique",
                "operationId": "06dbe76c8be83286e92a46b95c9c6f25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du marchand",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marchand trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marchand"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marchand non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Marchands"
                ],
                "summary": "Mettre à jour un marchand",
                "operationId": "4908ab9b4828806c29f7448c5438c09d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du marchand",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Marchand"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marchand mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marchand"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marchand non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Marchands"
                ],
                "summary": "Supprimer un marchand",
                "operationId": "a9e2d4ef00c2cd66d6f8563abaa7eef9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du marchand",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Marchand supprimé avec succès"
                    },
                    "404": {
                        "description": "Marchand non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Marchand": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "telephone": {
                        "type": "string"
                    },
                    "code_marchand": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    },
...
}